\name{LSDinterface-package}
\alias{LSDinterface-package}
\alias{LSDinterface}
\docType{package}
\title{
\packageTitle{LSDinterface}
}
\description{
\packageDescription{LSDinterface}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LSDinterface}
\packageIndices{LSDinterface}
LSD is free software copyrighted by Marco Valente and distributed under GNU  General Public License.
}
\author{
\packageAuthor{LSDinterface}

Maintainer: \packageMaintainer{LSDinterface}
}
\references{
LSD documentation is available at http://labsimdev.org and the latest binaries and source code can be downloaded at https://github.com/marcov64/Lsd.
}
\keyword{ package }
